+++ chain/src/store.rs
@@ -21,7 +21,7 @@ use codec::ChainCodec;
 use core::core::hash::{Hash, Hashed};
 use core::core::{Block, BlockHeader, Output};
 use grin_store::{self, Error, to_key, u64_to_key, option_to_not_found};
-use grin_store::codec::BlockCodec;
+use core::codec::BlockCodec;
 
 const STORE_SUBPATH: &'static str = "chain";
