+++ core/Cargo.toml
@@ -7,11 +7,13 @@ workspace = ".."
 [dependencies]
 bitflags = "~0.7.0"
 byteorder = "^0.5"
+bytes = "0.4.2"
 num-bigint = "^0.1.35"
 rust-crypto = "^0.2"
 rand = "^0.3"
 serde = "~0.9.10"
 serde_derive = "~0.9.10"
+tokio-io = "0.1.1"
 time = "^0.1"
 tiny-keccak = "1.1"
